DBZ-2982 Fix LogMinerMetrics bean registration

+++ debezium-connector-oracle/src/main/java/io/debezium/connector/oracle/logminer/LogMinerMetricsMXBean.java
@@ -97,7 +97,7 @@ public interface LogMinerMetricsMXBean {
      * sets number of milliseconds for connector to sleep before fetching another batch from the LogMiner view
      * @param milliseconds to sleep
      */
-    void setMillisecondToSleepBetweenMiningQuery(Integer milliseconds);
+    void setMillisecondToSleepBetweenMiningQuery(long milliseconds);
 
     /**
      * change sleeping time
